Aligned Images
I took the chess pieces from naraku9333's posting, and aligned them to
each have 80x80 pixel sqaure (making a 480x160 image).  I then made a
new board that uses this size, 80x80, for each square.  The helper is to
make it easier to align pieces in case that the positions should be
changed.  Note that the white and black pieces should be moved at the
same time.